+++ AlgoritmoTorreHanoiRecursivo.java
@@ -32,7 +32,7 @@ class AlgoritmoTorreHanoiRecursivo {
 	// Função recursiva que calcula o número de passos para resolver o problema, dado 
 	// a quantidade de peças.
 
-	public static int calcularPassos(int quantidadePecas){
+	public static long calcularPassos(long quantidadePecas){
 
 		if(quantidadePecas == 1){
 
@@ -66,7 +66,7 @@ class AlgoritmoTorreHanoiRecursivo {
 	public static void main(String[] args){
 
 		// Variável que recebe o número de peças
-		int quantidadePecas;
+		long quantidadePecas;
 
 		// Recebe a quantidade de peças e guarda na variável 'quantidadePecas'
 		System.out.println("Digite a quantidade de peças: ");
@@ -74,7 +74,7 @@ class AlgoritmoTorreHanoiRecursivo {
         quantidadePecas = in.nextInt();
 
         // Imprime o resultado no console
-		System.out.println("Número mínimo de passo para "+Integer.toString(quantidadePecas)+" peças: "+Integer.toString(calcularPassos(quantidadePecas)));
+		System.out.println("Número mínimo de passo para "+Long.toString(quantidadePecas)+" peças: "+Long.toString(calcularPassos(quantidadePecas)));
 		
 
 	}
